Updating scroll renderer to update and prepare for React implementation. Adding a md in the root to walk through implementation
